// storefront/src/web5/types.ts
export interface Store {
  url?: string; // URL of the store or marketplace
  headline?: string; // Headline for the store or marketplace
  description?: string; // Description for the store or marketplace
  provider: Partner // Provider of the store or marketplace
}

/* Uses partial properties from https://schema.org/Offer */
export interface Offer {
  inventoryLevel: string; // Inventory level of the associated Product
  /* Uses partial properties from https://schema.org/PriceSpecification */
  priceSpecification: {
    price: number; // Price for the Offer  
    priceCurrency: string; // Currency of the price for the associated Product 
  };
  seller: Partner; // Seller of the offer
  /* Uses partial properties from https://schema.org/Product */
  itemOffered: {
    name: string; // Name of the Product
    description: string; // Description of the Product
    /* Uses partial properties from https://schema.org/ImageObject */
    image: { 
      contentUrl: string; // URI of the image blob
      caption: string; // Alt text for the image
    },
  }
}

/* Uses partial properties from https://schema.org/AggregateOffer */
export interface AggregateOffer {
  /* Uses partial properties from https://schema.org/PriceSpecification */
  priceSpecification: {
    price: number; // Final price after markup   
    priceCurrency: string; // Currency of the price
  };
  /* Uses partial properties from https://schema.org/Offer */
  offers: {
    seller: Pick<Partner, "identifier">;
    identifier: string; // Fully qualified RecordID of the original Offer; ie {parentId}/{recordId}
  }[]; 
}
// ...
export interface Partner {
  identifier: string; // DID of the Partner
  name: string; // Name of the Partner
  email?: string; // Contact by email of the Partner
}